requirements.txt
gradio
psycopg2-binary
numpy
langchain
langchain-community
langchain-chroma
openai
langchain-openai
dbt-postgres 
prometheus_client
pytest
pytest-asyncio
pytest-mock
pytest-cov
python-docx
PyPDF2
python-magic